prep imdb data - modify get age function

+++ prepdata.py
@@ -3,14 +3,13 @@ import os, cv2
 import numpy as np
 from tqdm import tqdm
 
-if __name__ == '__main__':
-	RES_DIR = 'wiki_aligned'
-	data = loadfile.loadData('wiki', 'wiki.mat')
+def main(db_name):
+	RES_DIR = '{}_aligned'.format(db_name)
+	data = loadfile.loadData('{}'.format(db_name), '{}.mat'.format(db_name))
 	data = preprocess.cleanData(data)
-	data = data.iloc[:10]
+	# data = data.iloc[:10]
 	data = preprocess.loadImage(data)
 	print('[PREPROC] Run face alignment...')
-	# images = [preprocess.getAlignedFace(img) for img in tqdm(data['image'].tolist())]
 	paths = data['full_path'].tolist()
 	for (image, path) in tqdm(zip(data['image'].tolist(), paths), total=len(paths)):
 		image = preprocess.getAlignedFace(image)
@@ -18,4 +17,9 @@ if __name__ == '__main__':
 		if not os.path.exists(folder):
 			os.makedirs(folder)
 		flname = os.path.join(RES_DIR, path[0])
-		cv2.imwrite(flname, image)
+		if not os.path.exists(flname):
+			cv2.imwrite(flname, image)
+
+if __name__ == '__main__':
+	main('wiki')
+	main('imdb')
